+++ src/shared/group/channels.ts
@@ -2,6 +2,7 @@ export const GROUP_PUSH_CHANNEL = 'group:listChanged'
 
 export const GROUP_IPC = {
   list: 'group:list',
+  listLastActivity: 'group:listLastActivity',
   create: 'group:create',
   get: 'group:get',
   join: 'group:join',